Split the board building logic out
